페이지 요약 api 구현

+++ src/main/java/itopia/resolar/domain/page/Page.java
@@ -21,7 +21,7 @@ public class Page extends BaseEntity {
 
     private String url;
     private String summary;
-    private int importance;
+    private double importance;
 
     @ManyToOne(fetch = FetchType.LAZY)
     private Subject subject;
